+++ extensions/DSC/xBaLabWinSvrCfg.ps1
@@ -1,7 +1,7 @@
 ################################################################
 # Script to configure Windows lab environment using DSC        #
 # Author: Chris Langford                                       #
-# Version: 2.9.0                                               #
+# Version: 3.0.0                                               #
 ################################################################
 
 Configuration xBaMobilityandDevicesLabCfg {
@@ -477,7 +477,7 @@ Configuration xBaItEssentialsLabCfg {
     
 }
 
-Configuration xBaSuppItArchLabCfg {
+Configuration xBaSuppItArchP1LabCfg {
     [CmdletBinding()]
 
     Param (
@@ -552,14 +552,12 @@ Configuration xBaSuppItArchLabCfg {
         xScript "RunCreateVms"
         {
             SetScript = { 
-                New-VM -Name "Server01" -MemoryStartupBytes 1GB -Generation 2 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\Server01.vhdx" -SwitchName "vSwitch"
-                New-VM -Name "AsimLaptop" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\AsimLaptop.vhdx" -SwitchName "vSwitch"
-                New-VM -Name "ApprenticeLaptop" -MemoryStartupBytes 2GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\ApprenticeLaptop.vhdx" -SwitchName "vSwitch"
-                New-VM -Name "CallumPC" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\CallumPC.vhdx" -SwitchName "vSwitch"
-                New-VM -Name "KiaraLaptop" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\KiaraLaptop.vhdx" -SwitchName "vSwitch"
-                New-VM -Name "MariaPC" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\MariaPC.vhdx" -SwitchName "vSwitch"
-                New-VM -Name "SoniaPC" -MemoryStartupBytes 2GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\SoniaPC.vhdx" -SwitchName "vSwitch"
-                New-VM -Name "AdamLaptop" -MemoryStartupBytes 512MB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\AdamLaptop.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "Server01" -MemoryStartupBytes 1GB -Generation 2 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\Server01.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "AsimLaptop" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\AsimLaptop.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "ApprenticeLaptop" -MemoryStartupBytes 2GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\ApprenticeLaptop.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "CallumPC" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\CallumPC.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "KiaraLaptop" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\KiaraLaptop.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "MariaPC" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P1\MariaPC.vhdx" -SwitchName "vSwitch"
             }
             TestScript = { $false }
             GetScript = {  
@@ -597,7 +595,121 @@ Configuration xBaSuppItArchLabCfg {
     
 }
 
-Configuration xBaTrblshootNetP2LabCfg {
+Configuration xBaSuppItArchP2LabCfg {
+    [CmdletBinding()]
+
+    Param (
+        
+        [Parameter(Mandatory = $true)]
+        [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
+        $Credential
+    )
+
+    Import-DscResource -ModuleName xPSDesiredStateConfiguration
+
+    $features = @("Hyper-V", "RSAT-Hyper-V-Tools", "Hyper-V-Tools", "Hyper-V-PowerShell")
+
+    Node localhost {
+
+        LocalConfigurationManager {
+            RebootNodeIfNeeded = $true
+        }
+
+        # This resource block create a local user
+        xUser "CreateUserAccount" {
+            Ensure = "Present"
+            Username = Split-Path -Path $Credential.Username -Leaf
+            Password = $Credential
+            FullName = "Baltic Apprentice"
+            Description = "Baltic Apprentice"
+            PasswordNeverExpires = $true
+            PasswordChangeRequired = $false
+            PasswordChangeNotAllowed = $true
+        }
+
+        # This resource block adds user to a spacific group
+        xGroup "AddToRemoteDesktopUserGroup"
+        {
+            GroupName = "Remote Desktop Users"
+            Ensure = "Present"
+            MembersToInclude = Split-Path -Path $Credential.Username -Leaf
+            DependsOn = "[xUser]CreateUserAccount"
+        }
+
+        # This resource block adds user to a spacific group
+        xGroup "AddToHyperVAdministratorGroup"
+        {
+            GroupName = "Hyper-V Administrators"
+            Ensure = "Present"
+            MembersToInclude = Split-Path -Path $Credential.Username -Leaf
+            DependsOn = "[xUser]CreateUserAccount"
+        }
+
+        # This resource block ensures that a Windows Features (Roles) is present
+        xWindowsFeatureSet "AddHyperVFeatures"
+        {
+            Name = $features
+            Ensure = "Present"
+            IncludeAllSubFeature = $true
+        }
+
+        # This resource block ensures that the file is executed
+        xScript "RunvSwitchForNestedVms"
+        {
+            SetScript = { 
+                New-VMSwitch -SwitchName "vSwitch" -SwitchType Private
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+            DependsOn = "[xWindowsFeatureSet]AddHyperVFeatures"
+        }
+
+        xScript "RunCreateVms"
+        {
+            SetScript = {               
+                New-VM -Name "SoniaPC" -MemoryStartupBytes 2GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\SoniaPC.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "AdamLaptop" -MemoryStartupBytes 512MB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\SupportITArchitecture-P2\AdamLaptop.vhdx" -SwitchName "vSwitch"
+            }
+            TestScript = { $false }
+            GetScript = {  
+                # Do Nothing
+            }
+            DependsOn = "[xWindowsFeatureSet]AddHyperVFeatures"
+        }
+
+        # This resource block ensures that the file or command is executed
+        xScript "SetRdpTimeZone"
+        {
+            SetScript = {
+                New-ItemProperty -ErrorAction SilentlyContinue -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows NT\Terminal Services" -Name "fEnableTimeZoneRedirection" -Value "1" -PropertyType DWORD -Force
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+        }
+        
+        # This resource block ensures that the file or command is executed
+        xScript "RemoveArtifacts"
+        {
+            SetScript = {
+                Remove-Item "C:\workflow-artifacts\*" -Recurse -Force
+                Remove-Item "C:\workflow-artifacts" -Force
+                Remove-Item "C:\workflow-artifacts.zip" -Force
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+        }
+    }
+    
+}
+
+Configuration xBaTrblshootNetLabCfg {
     [CmdletBinding()]
 
     Param (
@@ -698,6 +810,223 @@ Configuration xBaTrblshootNetP2LabCfg {
         {
             SetScript = {
                 New-ItemProperty -ErrorAction SilentlyContinue -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows NT\Terminal Services" -Name "fEnableTimeZoneRedirection" -Value "1" -PropertyType DWORD -Force
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+        }
+
+        # This resource block ensures that the file or command is executed
+        xScript "RemoveArtifacts"
+        {
+            SetScript = {
+                Remove-Item "C:\workflow-artifacts\*" -Recurse -Force
+                Remove-Item "C:\workflow-artifacts" -Force
+                Remove-Item "C:\workflow-artifacts.zip" -Force
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+        }
+    }
+    
+}
+
+Configuration xBaEndUserDevicesP1LabCfg {
+    [CmdletBinding()]
+
+    Param (
+        
+        [Parameter(Mandatory = $true)]
+        [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
+        $Credential
+    )
+
+    Import-DscResource -ModuleName xPSDesiredStateConfiguration
+
+    $features = @("Hyper-V", "RSAT-Hyper-V-Tools", "Hyper-V-Tools", "Hyper-V-PowerShell")
+
+    Node localhost {
+
+        LocalConfigurationManager {
+            RebootNodeIfNeeded = $true
+        }
+
+        # This resource block create a local user
+        xUser "CreateUserAccount" {
+            Ensure = "Present"
+            Username = Split-Path -Path $Credential.Username -Leaf
+            Password = $Credential
+            FullName = "Baltic Apprentice"
+            Description = "Baltic Apprentice"
+            PasswordNeverExpires = $true
+            PasswordChangeRequired = $false
+            PasswordChangeNotAllowed = $true
+        }
+
+        # This resource block adds user to a spacific group
+        xGroup "AddToRemoteDesktopUserGroup"
+        {
+            GroupName = "Remote Desktop Users"
+            Ensure = "Present"
+            MembersToInclude = Split-Path -Path $Credential.Username -Leaf
+            DependsOn = "[xUser]CreateUserAccount"
+        }
+
+        # This resource block adds user to a spacific group
+        xGroup "AddToHyperVAdministratorGroup"
+        {
+            GroupName = "Hyper-V Administrators"
+            Ensure = "Present"
+            MembersToInclude = Split-Path -Path $Credential.Username -Leaf
+            DependsOn = "[xUser]CreateUserAccount"
+        }
+
+        # This resource block ensures that a Windows Features (Roles) is present
+        xWindowsFeatureSet "AddHyperVFeatures"
+        {
+            Name = $features
+            Ensure = "Present"
+            IncludeAllSubFeature = $true
+        }        
+
+        # This resource block ensures that the file or command is executed
+        xScript "SetRdpTimeZone"
+        {
+            SetScript = {
+                New-ItemProperty -ErrorAction SilentlyContinue -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows NT\Terminal Services" -Name "fEnableTimeZoneRedirection" -Value "1" -PropertyType DWORD -Force
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+        }
+
+        # This resource block ensures that the file or command is executed
+        xScript "RemoveArtifacts"
+        {
+            SetScript = {
+                Remove-Item "C:\workflow-artifacts\*" -Recurse -Force
+                Remove-Item "C:\workflow-artifacts" -Force
+                Remove-Item "C:\workflow-artifacts.zip" -Force
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+        }
+    }
+    
+}
+
+Configuration xBaEndUserDevicesP2LabCfg {
+    [CmdletBinding()]
+
+    Param (
+        
+        [Parameter(Mandatory = $true)]
+        [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
+        $Credential
+    )
+
+    Import-DscResource -ModuleName xPSDesiredStateConfiguration
+
+    $features = @("Hyper-V", "RSAT-Hyper-V-Tools", "Hyper-V-Tools", "Hyper-V-PowerShell")
+
+    Node localhost {
+
+        LocalConfigurationManager {
+            RebootNodeIfNeeded = $true
+        }
+
+        # This resource block create a local user
+        xUser "CreateUserAccount" {
+            Ensure = "Present"
+            Username = Split-Path -Path $Credential.Username -Leaf
+            Password = $Credential
+            FullName = "Baltic Apprentice"
+            Description = "Baltic Apprentice"
+            PasswordNeverExpires = $true
+            PasswordChangeRequired = $false
+            PasswordChangeNotAllowed = $true
+        }
+
+        # This resource block adds user to a spacific group
+        xGroup "AddToRemoteDesktopUserGroup"
+        {
+            GroupName = "Remote Desktop Users"
+            Ensure = "Present"
+            MembersToInclude = Split-Path -Path $Credential.Username -Leaf
+            DependsOn = "[xUser]CreateUserAccount"
+        }
+
+        # This resource block adds user to a spacific group
+        xGroup "AddToHyperVAdministratorGroup"
+        {
+            GroupName = "Hyper-V Administrators"
+            Ensure = "Present"
+            MembersToInclude = Split-Path -Path $Credential.Username -Leaf
+            DependsOn = "[xUser]CreateUserAccount"
+        }
+
+        # This resource block ensures that a Windows Features (Roles) is present
+        xWindowsFeatureSet "AddHyperVFeatures"
+        {
+            Name = $features
+            Ensure = "Present"
+            IncludeAllSubFeature = $true
+        }
+
+        # This resource block ensures that the file is executed
+        xScript "RunvSwitchForNestedVms"
+        {
+            SetScript = { 
+                New-VMSwitch -SwitchName "vSwitch" -SwitchType Private
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+            DependsOn = "[xWindowsFeatureSet]AddHyperVFeatures"
+        }
+
+        xScript "RunCreateVms"
+        {
+            SetScript = { 
+                New-VM -Name "Server01" -MemoryStartupBytes 2GB -Generation 2 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\WWEndUserDevices\WWEUD-Server01.vhdx" -SwitchName "vSwitch"
+                New-VM -Name "WindowsClient" -MemoryStartupBytes 1GB -Generation 1 -BootDevice VHD -VHDPath "C:\Users\Public\Documents\Hyper-V\Virtual hard disks\WWEndUserDevices\WWEUD-Win10.vhdx" -SwitchName "vSwitch"
+                
+            }
+            TestScript = { $false }
+            GetScript = {  
+                # Do Nothing
+            }
+            DependsOn = "[xWindowsFeatureSet]AddHyperVFeatures"
+        }
+
+        # This resource block ensures that the file or command is executed
+        xScript "SetRdpTimeZone"
+        {
+            SetScript = {
+                New-ItemProperty -ErrorAction SilentlyContinue -Path "HKLM:\SOFTWARE\Policies\Microsoft\Windows NT\Terminal Services" -Name "fEnableTimeZoneRedirection" -Value "1" -PropertyType DWORD -Force
+            }
+            TestScript = { $false }
+            GetScript = { 
+                # Do Nothing
+            }
+        }
+
+        # This resource block ensures that the file or command is executed
+        xScript "RemoveArtifacts"
+        {
+            SetScript = {
+                Remove-Item "C:\workflow-artifacts\*" -Recurse -Force
+                Remove-Item "C:\workflow-artifacts" -Force
+                Remove-Item "C:\workflow-artifacts.zip" -Force
             }
             TestScript = { $false }
             GetScript = {